create math module. move jd_to_mjd and mag_to_flux in. update supernnova
