+++ broker/cloud_functions/classify_snn/main.py
@@ -12,7 +12,7 @@ import pandas as pd
 from pathlib import Path
 from supernnova.validation.validate_onthefly import classify_lcs
 
-from broker_utils import data_utils, gcp_utils, schema_maps
+from broker_utils import data_utils, gcp_utils, math, schema_maps
 
 
 PROJECT_ID = os.getenv("GCP_PROJECT")
@@ -125,8 +125,8 @@ def _format_for_snn(alert_dict: dict) -> pd.DataFrame:
     snn_df["FLT"] = alert_df["fid"].map(schema_map["FILTER_MAP"])
 
     if SURVEY == "ztf":
-        snn_df["MJD"] = data_utils.jd_to_mjd(alert_df["jd"])
-        snn_df["FLUXCAL"], snn_df["FLUXCALERR"] = data_utils.mag_to_flux(
+        snn_df["MJD"] = math.jd_to_mjd(alert_df["jd"])
+        snn_df["FLUXCAL"], snn_df["FLUXCALERR"] = math.mag_to_flux(
             alert_df[schema_map["mag"]],
             alert_df[schema_map["magzp"]],
             alert_df[schema_map["magerr"]],
